+++ src/model/element-base.ts
@@ -19,6 +19,7 @@ import * as jsdoc from '../javascript/jsdoc';
 import {Class, ClassInit} from './class';
 import {Privacy} from './feature';
 import {Attribute, Document, Event, Feature, Method, Resolvable, ScannedAttribute, ScannedEvent, ScannedProperty, ScannedReference, SourceRange, Warning} from './model';
+import {Severity} from './warning';
 
 export {Visitor} from '../javascript/estree-visitor';
 
@@ -44,7 +45,27 @@ export abstract class ScannedElementBase implements Resolvable {
   superClass?: ScannedReference = undefined;
 
   applyHtmlComment(commentText: string|undefined) {
-    this.description = this.description || commentText || '';
+    if (commentText) {
+      const commentJsdoc = jsdoc.parseJsdoc(commentText);
+      // Add a Warning if there are already jsdoc tags or a description for this
+      // element.
+      if (this.sourceRange &&
+          (this.description || this.jsdoc && this.jsdoc.tags.length > 0)) {
+        this.warnings.push({
+          severity: Severity.WARNING,
+          code: 'multiple-doc-comments',
+          message:
+              `${this.constructor.name} has both HTML doc and JSDoc comments.`,
+          sourceRange: this.sourceRange,
+        });
+      }
+      this.jsdoc =
+          this.jsdoc ? jsdoc.join(commentJsdoc, this.jsdoc) : commentJsdoc;
+      this.description = [
+        commentJsdoc.description || '',
+        this.description || ''
+      ].join('\n\n').trim();
+    }
   }
 
   applyJsdocDemoTags(baseUrl: string): void {